[FIX] warnings removed
git-svn-id: http://svn.seqan.de/seqan/trunk@12761 e6417c60-b987-48fd-844e-b20f0fcc1017

+++ seqan/extras/apps/razers3/razers_match_filter.h
@@ -146,7 +146,7 @@ _newLimitDistRange(TMatchFilter const & filter, unsigned histogramId)
         while (it != itEnd && *it == 0)
             ++it;
 
-        return _min((it - itBeg) + filter.options.scoreDistanceRange, itEnd - itBeg);
+        return _min((it - itBeg) + (int)filter.options.scoreDistanceRange, itEnd - itBeg);
     }
 
     // we need at least one more than max before we can purge
